derived: renamed derived_measurments --> energy
Renamed to reduce redandancy in import path.
